backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
python-dotenv
cloudinary
psycopg2-binary
pytest
pytest-asyncio
pillow
easyocr
deepl
